Testing Multiprocessing - See Comments  Editing Function Code to save into dict manger I am testiing whether keys of the list need to be defined in the function.
Right in the begining:

"But that might get inefficient if the list is going to grow long - the whole list will be serialised and sent to the manager process for each append. A better way in that case would be to make the shared list directly with SyncManager.list (although you'll still have the same problem if the elements of the list are mutable - you need to reset them in the list to send them between the processes)."

Little bit more testing, and I will be back on the course.

http://stackoverflow.com/questions/35202278/cannot-append-items-to-multiprocessing-shared-list