Functional Design of Linear Regression - Read Comments
For the Test Sample: I believe if I apply K-nearest Neighbour to the sub features and apply linear regression to the entire sample. I should be receiving a decent score on the performance of the prediction.
Not sure at the moment.
I still have to read the paper in detial some other time.
Just noticed the bottom references after going through some parts of it.